+++ pages/_document.tsx
@@ -5,6 +5,11 @@ export default function Document() {
     <Html lang="ko">
       <Head />
       <body>
+        <script
+          src="https://t1.kakaocdn.net/kakao_js_sdk/2.1.0/kakao.min.js"
+          integrity="sha384-dpu02ieKC6NUeKFoGMOKz6102CLEWi9+5RQjWSV0ikYSFFd8M3Wp2reIcquJOemx"
+          crossOrigin="anonymous"
+        ></script>
         <Main />
         <NextScript />
       </body>
